Use mostRecentInPlayId
