+++ server/game/stateWatchers/CardsDefeatedThisPhaseWatcher.ts
@@ -1,5 +1,4 @@
 import { StateWatcher } from '../core/stateWatcher/StateWatcher';
-import type { CardType, Trait, ZoneName } from '../core/Constants';
 import { StateWatcherName } from '../core/Constants';
 import type { StateWatcherRegistrar } from '../core/stateWatcher/StateWatcherRegistrar';
 import type { Player } from '../core/Player';
@@ -10,18 +9,7 @@ import type { Game } from '../core/Game';
 import type { UnwrapRef, UnwrapRefObject } from '../core/GameObjectBase';
 import type { IDefeatSource } from '../IDamageOrDefeatSource';
 import { registerState, type GameObjectId } from '../core/GameObjectUtils';
-
-
-/**
- * Simplified last known information for defeated cards
- */
-export interface IStateWatcherLKIEntry {
-    traits: Set<Trait>;
-    type: CardType;
-    power?: number;
-    arena?: ZoneName;
-    // TODO: Add more fields if needed
-}
+import type { IStateWatcherLKIEntry } from '../core/stateWatcher/StateWatcher';
 
 export interface DefeatedCardEntry {
     card: GameObjectId<IInPlayCard>;
